Moved draft sources utility into the draft sources service
This was necessary to be able to mock out the utility method. It will make unit testing much, much easier.
