+++ src/SIL.XForge.Scripture/ClientApp/src/app/translate/draft-generation/draft-sources.service.ts
@@ -1,5 +1,4 @@
 import { Injectable } from '@angular/core';
-import { WritingSystem } from 'realtime-server/lib/esm/common/models/writing-system';
 import { TextInfo } from 'realtime-server/lib/esm/scriptureforge/models/text-info';
 import { TranslateConfig, TranslateSource } from 'realtime-server/lib/esm/scriptureforge/models/translate-config';
 import { combineLatest, defer, from, Observable } from 'rxjs';
@@ -13,12 +12,8 @@ import { SFProjectService } from '../../core/sf-project.service';
 interface DraftTextInfo {
   bookNum: number;
 }
-export interface DraftSource {
-  name: string;
-  projectRef: string;
-  shortName: string;
+export interface DraftSource extends TranslateSource {
   texts: DraftTextInfo[];
-  writingSystem: WritingSystem;
   noAccess?: boolean;
 }
 interface DraftSourceDoc {
@@ -31,6 +26,12 @@ export interface DraftSources {
   alternateSource?: Readonly<DraftSource>;
   alternateTrainingSource?: Readonly<DraftSource>;
   additionalTrainingSource?: Readonly<DraftSource>;
+}
+
+export interface TranslateSourcesAsArrays {
+  trainingSources: [TranslateSource?, TranslateSource?];
+  trainingTargets: [TranslateSource];
+  draftingSources: [TranslateSource?];
 }
 
 @Injectable({
@@ -157,6 +158,7 @@ export class DraftSourcesService {
         data: {
           name: translateSource.name,
           shortName: translateSource.shortName,
+          paratextId: translateSource.paratextId,
           projectRef: projectId,
           texts: texts?.slice() ?? [],
           writingSystem: translateSource.writingSystem,
@@ -167,5 +169,61 @@ export class DraftSourcesService {
       // The real document will be read from the realtime server, as the current user has permission
       return undefined;
     }
+  }
+
+  /**
+   * Returns the training and drafting sources for the current project as three arrays.
+   *
+   * This considers properties such as alternateTrainingSourceEnabled and alternateTrainingSource and makes sure to only
+   * include a source if it's enabled and not null. It also considers whether the project source is implicitly the
+   * training and/or drafting source.
+   *
+   * This method is also intended to be act as an abstraction layer to allow changing the data model in the future
+   * without needing to change all the places that use this method.
+   *
+   * Currently this method provides guarantees via the type system that there will be at most 2 training sources,
+   * exactly 1 training target, and at most 1 drafting source. Consumers of this method that cannot accept an arbitrary
+   * length for each of these arrays are encouraged to write there code in such a way that it will noticeably break
+   * (preferably at build time) if these guarantees are changed, to make it easier to find code that relies on the
+   * current limit on the number of sources in each category.
+   * @param target The project to get the sources for
+   * @param targetId The projectRef of the target
+   * @returns An object with three arrays: trainingSources, trainingTargets, and draftingSources
+   */
+  getTranslateSources(): TranslateSourcesAsArrays {
+    const target = this.activatedProject.projectDoc.data;
+    const trainingSources: [TranslateSource?, TranslateSource?] = [];
+    const draftingSources: [TranslateSource?] = [];
+    const trainingTargets: [TranslateSource] = [{ ...target, projectRef: this.activatedProject.projectId }];
+
+    const draftConfig = target.translateConfig.draftConfig;
+
+    let trainingSource: TranslateSource | undefined;
+    if (draftConfig.alternateTrainingSourceEnabled && draftConfig.alternateTrainingSource != null) {
+      trainingSource = draftConfig.alternateTrainingSource;
+    } else {
+      trainingSource = target.translateConfig.source;
+    }
+
+    if (trainingSource != null) {
+      trainingSources.push(trainingSource);
+    }
+
+    if (draftConfig.additionalTrainingSourceEnabled && draftConfig.additionalTrainingSource != null) {
+      trainingSources.push(draftConfig.additionalTrainingSource);
+    }
+
+    let draftingSource: TranslateSource | undefined;
+    if (draftConfig.alternateSourceEnabled && draftConfig.alternateSource != null) {
+      draftingSource = draftConfig.alternateSource;
+    } else {
+      draftingSource = target.translateConfig.source;
+    }
+
+    if (draftingSource != null) {
+      draftingSources.push(draftingSource);
+    }
+
+    return { trainingSources, trainingTargets, draftingSources };
   }
 }
